properties: modernize the UI file
The ultimate goal is to eventually get rid of stuff (widget classes and
properties) that's no longer supported in Gtk4 for ease of porting.

This gets rid of GtkAlignment and GtkTable: The former is replaced by
doing the alignment and spacing using properties of the widgets themselves
or their parents. The style is changed a bit to match whatu
gnome-control-center and recent versions of nm-connection-editor use.
The later is replaced with GtkGrid.

The patch lumps a lot of changes together and is generally
incomprehensible because it's XML. Such is life.